Rename "personalization" to "appearance"
The difference between Preference and personalization
is not obvious, so the modify the name.

TODO: Update translations other than English and Chinese.
