+++ src/ui/flutter_app/lib/l10n/intl_da.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalisering",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalisering",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Udnyt den menneskelige erfaring",
   "@drawOnHumanExperience": {
