+++ src/ui/flutter_app/lib/l10n/intl_de_ch.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalisierige",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalisierige",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Uf Mönschlechi Erfahrige zrüggrife",
   "@drawOnHumanExperience": {
