+++ src/ui/flutter_app/lib/l10n/intl_es.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalización",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalización",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Basarse en la experiencia humana",
   "@drawOnHumanExperience": {
