+++ src/ui/flutter_app/lib/l10n/intl_fa.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "تنظیمات ظاهر",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "تنظیمات ظاهر",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "استفاده از تجربه انسان",
   "@drawOnHumanExperience": {
