+++ src/ui/flutter_app/lib/l10n/intl_fr.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personnalisation",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personnalisation",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "S'inspirer de l'expérience humaine",
   "@drawOnHumanExperience": {
