+++ src/ui/flutter_app/lib/l10n/intl_gu.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "વૈયક્તિકરણ",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "વૈયક્તિકરણ",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "માનવ અનુભવ દોરો",
   "@drawOnHumanExperience": {
