+++ src/ui/flutter_app/lib/l10n/intl_hr.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalizacija",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalizacija",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Oslonite se na ljudsko iskustvo",
   "@drawOnHumanExperience": {
