+++ src/ui/flutter_app/lib/l10n/intl_it.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalizzazione",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalizzazione",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Attingere all'esperienza umana",
   "@drawOnHumanExperience": {
