+++ src/ui/flutter_app/lib/l10n/intl_ja.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "外観設定",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "外観設定",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "人間の経験を借りる",
   "@drawOnHumanExperience": {
