+++ src/ui/flutter_app/lib/l10n/intl_lt.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalizavimas",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalizavimas",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Remkitės žmogaus patirtimi",
   "@drawOnHumanExperience": {
