+++ src/ui/flutter_app/lib/l10n/intl_mk.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Персонализација",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Персонализација",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Извлечете го човечкото искуство",
   "@drawOnHumanExperience": {
