+++ src/ui/flutter_app/lib/l10n/intl_ms.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Pemperibadian",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Pemperibadian",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Ikuti pengalaman manusia",
   "@drawOnHumanExperience": {
