+++ src/ui/flutter_app/lib/l10n/intl_nb.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalisering",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalisering",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Tegn på den menneskelige opplevelsen",
   "@drawOnHumanExperience": {
