+++ src/ui/flutter_app/lib/l10n/intl_nl.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalisering",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalisering",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Maak gebruik van de menselijke ervaring",
   "@drawOnHumanExperience": {
