+++ src/ui/flutter_app/lib/l10n/intl_ru.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Персонализация",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Персонализация",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Опирайтесь на человеческий опыт",
   "@drawOnHumanExperience": {
