+++ src/ui/flutter_app/lib/l10n/intl_sr.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Персонализација",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Персонализација",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Ослоните се на људско искуство",
   "@drawOnHumanExperience": {
