+++ src/ui/flutter_app/lib/l10n/intl_sv.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Personalisering",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Personalisering",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Dra nytta av den mänskliga erfarenheten",
   "@drawOnHumanExperience": {
