+++ src/ui/flutter_app/lib/l10n/intl_uz.arb
@@ -777,9 +777,9 @@
   "@timeout": {
     "description": "Timeout"
   },
-  "personalization": "Shaxsiylashtirish",
-  "@personalization": {
-    "description": "Personalization"
+  "appearance": "Shaxsiylashtirish",
+  "@appearance": {
+    "description": "Appearance"
   },
   "drawOnHumanExperience": "Inson tajribasidan foydalaning",
   "@drawOnHumanExperience": {
